+++ dso/branches/santiago_lock_rewrite/code/base/dso-l1-api/src/com/tc/object/bytecode/TransparencyClassAdapter.java
@@ -19,7 +19,7 @@ import com.tc.object.Portability;
 import com.tc.object.config.ConfigLockLevel;
 import com.tc.object.config.LockDefinition;
 import com.tc.object.config.TransparencyClassSpec;
-import com.tc.object.lockmanager.api.LockLevel;
+import com.tc.object.locks.LockLevel;
 import com.tc.object.logging.InstrumentationLogger;
 import com.tc.text.Banner;
 import com.tc.util.Assert;
@@ -197,7 +197,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
           instrumentationLogger.autolockInserted(this.spec.getClassNameDots(), name, desc, ld);
         }
       }
-      boolean isAutoReadLock = isAutolock && (lockLevel == LockLevel.READ);
+      boolean isAutoReadLock = isAutolock && (lockLevel == LockLevel.READ.toInt());
 
       if (isAutoSynchronized(ld) && !"<init>".equals(name)) {
         access |= ACC_SYNCHRONIZED;
@@ -590,7 +590,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       gv.visitTryCatchBlock(l0, l1, l2, null);
       gv.visitLabel(l0);
 
-      callVolatileBegin(spec.getClassNameDots() + "." + name, LockLevel.READ, gv);
+      callVolatileBegin(spec.getClassNameDots() + "." + name, LockLevel.READ.toInt(), gv);
 
       Label l6 = new Label();
       gv.visitJumpInsn(JSR, l6);
@@ -606,7 +606,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       gv.visitLabel(l6);
       gv.visitVarInsn(ASTORE, 1);
 
-      callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ, gv);
+      callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ.toInt(), gv);
       gv.visitVarInsn(RET, 1);
     }
 
@@ -694,7 +694,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
         }
       }
 
-      callTCBeginWithLockName(rootName, LockLevel.WRITE, mv);
+      callTCBeginWithLockName(rootName, LockLevel.WRITE.toInt(), mv);
 
       mv.visitLabel(l3);
       mv.visitLdcInsn(rootName);
@@ -718,7 +718,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
 
       mv.visitLabel(l8);
       mv.visitVarInsn(ASTORE, 2);
-      callTCCommitWithLockName(rootName, LockLevel.WRITE, mv);
+      callTCCommitWithLockName(rootName, LockLevel.WRITE.toInt(), mv);
       mv.visitVarInsn(RET, 2);
 
       mv.visitLabel(l5);
@@ -759,7 +759,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       gv.visitJumpInsn(IFNULL, l0);
 
       if (isVolatile) {
-        callVolatileBegin(spec.getClassNameDots() + '.' + name, LockLevel.READ, gv);
+        callVolatileBegin(spec.getClassNameDots() + '.' + name, LockLevel.READ.toInt(), gv);
       }
 
       gv.visitVarInsn(ALOAD, 3);
@@ -788,7 +788,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
 
       gv.visitInsn(MONITOREXIT);
       if (isVolatile) {
-        callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ, gv);
+        callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ.toInt(), gv);
       }
       gv.visitInsn(ARETURN);
       gv.visitJumpInsn(GOTO, l0);
@@ -798,7 +798,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
 
       gv.visitInsn(MONITOREXIT);
       if (isVolatile) {
-        callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ, gv);
+        callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.READ.toInt(), gv);
       }
       gv.visitInsn(ATHROW);
 
@@ -845,7 +845,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       scv.visitTryCatchBlock(l0, l1, l2, null);
       scv.visitLabel(l0);
 
-      callVolatileBegin(spec.getClassNameDots() + "." + name, LockLevel.WRITE, scv);
+      callVolatileBegin(spec.getClassNameDots() + "." + name, LockLevel.WRITE.toInt(), scv);
 
       Label l6 = new Label();
       scv.visitJumpInsn(JSR, l6);
@@ -862,7 +862,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       scv.visitInsn(ATHROW);
       scv.visitLabel(l6);
       scv.visitVarInsn(ASTORE, 1);
-      callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.WRITE, scv);
+      callVolatileCommit(spec.getClassNameDots() + "." + name, LockLevel.WRITE.toInt(), scv);
 
       scv.visitVarInsn(RET, 1);
     }
@@ -945,7 +945,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       }
 
       String rootName = rootNameFor(spec.getClassNameSlashes(), name);
-      callTCBeginWithLockName(rootName, LockLevel.WRITE, scv);
+      callTCBeginWithLockName(rootName, LockLevel.WRITE.toInt(), scv);
 
       scv.visitLabel(tryStart);
 
@@ -981,7 +981,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
 
       scv.visitLabel(finallyStart);
       scv.visitVarInsn(ASTORE, 2);
-      callTCCommitWithLockName(rootName, LockLevel.WRITE, scv);
+      callTCCommitWithLockName(rootName, LockLevel.WRITE.toInt(), scv);
       scv.visitVarInsn(RET, 2);
 
       scv.visitLabel(normalExit);
@@ -1090,7 +1090,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       scv.visitJumpInsn(IFNULL, labelArgumentIsNull);
 
       if (isVolatile) {
-        generateCodeForVolatileTransactionBegin(l1, l2, labelArgumentIsNull, l4, spec.getClassNameDots() + "." + name, LockLevel.WRITE,
+        generateCodeForVolatileTransactionBegin(l1, l2, labelArgumentIsNull, l4, spec.getClassNameDots() + "." + name, LockLevel.WRITE.toInt(),
                                                 scv);
       }
 
@@ -1103,7 +1103,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
                           "(Ljava/lang/String;Ljava/lang/String;Ljava/lang/Object;I)V");
 
       if (isVolatile) {
-        generateCodeForVolativeTransactionCommit(l1, l2, scv, 3, 4, spec.getClassNameDots() + "." + name, LockLevel.WRITE);
+        generateCodeForVolativeTransactionCommit(l1, l2, scv, 3, 4, spec.getClassNameDots() + "." + name, LockLevel.WRITE.toInt());
       }
 
       scv.visitLabel(labelArgumentIsNull);
@@ -1142,7 +1142,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
       mv.visitJumpInsn(IFNULL, l0);
 
       if (isVolatile) {
-        generateCodeForVolatileTransactionBegin(l1, l2, l0, l4, spec.getClassNameDots() + "." + name, LockLevel.WRITE,
+        generateCodeForVolatileTransactionBegin(l1, l2, l0, l4, spec.getClassNameDots() + "." + name, LockLevel.WRITE.toInt(),
                                                 mv);
       }
 
@@ -1157,7 +1157,7 @@ public class TransparencyClassAdapter extends ClassAdapterBase {
 
       if (isVolatile) {
         generateCodeForVolativeTransactionCommit(l1, l2, mv, 2 + t.getSize(), 3 + t.getSize(), spec.getClassNameDots()
-                                                                                               + "." + name, LockLevel.WRITE);
+                                                                                               + "." + name, LockLevel.WRITE.toInt());
       }
 
       mv.visitLabel(l0);
